LogicGame

+++ Snake/core/src/sprites/Bodypart.java
@@ -3,11 +3,11 @@ package sprites;
 import com.badlogic.gdx.math.Vector3;
 
 public class Bodypart {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private Vector3 position;
     private Controls controls;
-    public Bodypart(int x, int y, int boardSize) {
+    public Bodypart(float x, float y, int boardSize) {
         position = new Vector3(x, y, 0);
         controls = new Controls();
         this.x = x% boardSize;
@@ -18,11 +18,11 @@ public class Bodypart {
 
     }
 
-    public int getX () {
+    public float getX () {
         return x;
     }
-    public int getY () {
-        return y;
+    public float getY () {
+        return  y;
     }
 }
